requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
redis>=5.0.0
playwright>=1.41.0
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.4.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
selectolax>=0.3.21
httpx>=0.26.0
python-dotenv>=1.0.1
arq>=0.26.0
